Working cross-compile build
This enables cross-compile in a single configure/make sequence. The hard part was that there's a variety of programs in patterns that get used at compile time, and so they need to be compiled for the build machine, but they use libraries from other folders that are also used by the host machine.

To solve this, the relevant folders have a lib<foo>-host.a build as well as a lib<foo>.a build. The latter is for the compile target, the former is for the host target. This is done in a pretty crude way, but seems to work so far.

Signed-off-by: Tom Parker-Shemilt <palfrey@tevp.net>